backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.3
protobuf==5.29.0
yfinance==0.2.59
pandas==2.2.0
numpy==1.26.4
tensorflow==2.18.1
scikit-learn==1.4.0
ta==0.11.0
pytest==7.4.4
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
alembic==1.13.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
schedule==1.2.1